+++ packages/react-hooks/CHANGELOG.md
@@ -1,8 +1,8 @@
-# @vmejs/react-hooks
+# @ifsjs/react-hooks
 
 ## 0.0.1-alpha.5
 
 ### Patch Changes
 
-- Updated dependencies [[`9c86f61`](https://github.com/vmejs/vmejs/commit/9c86f6108d4a1f96e3e85338ae6263fcb433a68d)]:
-  - @vmejs/shared@0.0.1
+- Updated dependencies [[`9c86f61`](https://github.com/scalerone/ifsjs/commit/9c86f6108d4a1f96e3e85338ae6263fcb433a68d)]:
+  - @ifsjs/shared@0.0.1