+++ playgrounds/react/CHANGELOG.md
@@ -1,3 +1,3 @@
-# @vmejs/react-playground
+# @ifsjs/react-playground
 
 ## 0.0.1